requirements.txt
pillow==11.0.0
pyaes==1.6.1
pyasn1==0.6.1
rsa==4.9
Telethon==1.38.1
